Extracted StudentResponse class out to Models.
